\name{RadialPlotter-package}
\alias{RadialPlotter-package}
\docType{package}
\title{Package for OSL equivalent dose analysis and radial plot drawing }
\description{
This package performs statistical age models (log-scale) analysis and radial plot drawing.}
\details{
\tabular{ll}{
Package: \tab RadialPlotter\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-20\cr
License: \tab GPL-3\cr
}
}
\author{

Peng Jun

\bold{Package maintainer}

Peng Jun
\email{pengjun10@mails.ucas.ac.cn}
}
\references{
Galbraith, R.F., 1988. Graphical Display of Estimates Having Differing Standard Errors. Technometrics, 30 (3), pp. 271-281.
}
\keyword{package}
