\name{RadialPlotter-package}
\alias{RadialPlotter-package}
\docType{package}
\title{Package for OSL equivalent dose analysis and radial plot drawing }
\description{
This package performs statistical age models (log-scale) analysis and radial plot drawing. Routines for statistical age models analysis using fortran subroutines wrapped function 'RadialPlotter'. Facilitating the analysis of Galbraith's statistical age models, that are: \cr
\cr
Central age model (CAM) \cr
Finite mixture age model (FMM) \cr
Minimum age model with three parameters (MAM3) \cr
Minimum age model with four parameters (MAM4) \cr
}
\details{
\tabular{ll}{
Package: \tab RadialPlotter\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2013-04-20\cr
License: \tab GPL-3\cr
}
}
\author{
\tabular{ll}{
Peng Jun \tab University of Chinese Academy of Sciences, Beijing, China
}
\bold{Package maintainer}

Peng Jun
\email{pengjun10@mails.ucas.ac.cn}
}
\references{
Galbraith, R.F., 1988. Graphical Display of Estimates Having Differing Standard Errors. Technometrics, 30 (3), pp. 271-281.
}
\keyword{package}
